Refactor views.py
