+++ calyx/src/courses/views/__init__.py
@@ -1,0 +1,4 @@
+from .courses import CourseViewSet, CourseConfigViewSet, YearViewSet
+from .course_user import CourseAdminView, JoinAPIView, RequirementStatusView
+from .labs import LabViewSet
+from .ranks import RankViewSet
